Rewrite to backend-frontend architecture
The original implementation is in the "single_binary_host" dir.

There are 2 new frontends: plain (hosted by Rust, but its not
necessary), and sveltekit (not great code, i'm not a frontend dev).

There is currently only 1 backend: "backend-fast". "fast" because it is
using my custom in-memory, commit on write DB implementation instead of SQL.

In the future I also plan to make "backend-scale" version which would
connect to separate DB via SQL which will allow to spawn multiple instances
behind a load balancer to spread the load for big deployments.

Stay tuned :-)